Change for toggling

+++ kill-script.sh
@@ -1,4 +1,8 @@
 #!/bin/bash
 
 read -a info
-kill ${info[1]}
+if [ "${info[2]}" == "0.0" ]; then
+        /usr/bin/gnome-calculator
+else
+        kill ${info[1]}
+fi
